Refactor UniqueList and Javadocs for consistency
At present, there may be confusion related to UniqueList with respect to
what is considered unique, especially when it changes according to
situation. Specifically, `set` and `remove` uses `Objects.equals` in
order to ensure the exact element is replaced or removed, while other
cases uses the `isDistinct` method to define uniqueness. The Javadocs
adds to the confusion by using a mixture of terminology like unique,
distinct, equivalent, interchangebly.

Let's
* Ensure terminology is consistent. We will use "the element", "matching
  element", "exact element", to refer to a match using `Objects.equals`.
  We will use "equivalent element", "unique element", to refer to a
  match using `isEquivalent`.
* Refactor UniqueList to ensure terminology is consistent. We refactor
  `isDistinct` to `isEquivalent`.
* Update existing confusing code - `UniqueAttendanceRecordList.remove`
  breaks expectations by checking for equivalence instead of the exact
  element. We fix this by removing the override and updating code that
  depends on the behaviour.
* Add a `find` method for UniqueList that finds based on equivalence to
  provide better support for using methods that require exact match,
  while doubling as a convenient search mechanism.